% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_api_post.R
\name{call_that_api_post}
\alias{call_that_api_post}
\title{Submits a POST command against a REST API}
\usage{
call_that_api_post(
  api_connection,
  endpoint = "",
  body = NULL,
  headers = list(),
  ...
)
}
\arguments{
\item{api_connection}{A \code{call_that_connection} object}

\item{endpoint}{The name of the API's endpoint.  Defaults to an empty character vector.}

\item{body}{The body or query sent to the REST API.}

\item{headers}{A \code{list} object containing a named list of headers to pass to the API.}

\item{...}{Other arguments to pass to the REST API call.}
}
\value{
An \code{httr} \code{request} object
}
\description{
Submits a POST command against a REST API
}
