% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_session.R
\name{call_that_available_tests}
\alias{call_that_available_tests}
\title{Matches APIs to tests}
\usage{
call_that_available_tests(
  test_directory = "tests/testthat",
  plumber_directory = "inst/plumber"
)
}
\arguments{
\item{test_directory}{Location of the test scripts. Defaults to 'test/testthat'.}

\item{plumber_directory}{Location of the plumber APIs. Defaults to 'inst/plumber'.}
}
\description{
Matches APIs to tests
}
\details{
It looks for test scripts with the prefix 'test-plumber-...'. It
matches the last part of the script's name and matches it to plumber API
inside the 'inst/plumber' folder.
}
