% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_connection.R
\name{call_that_connection}
\alias{call_that_connection}
\title{Creates a generic API connection object}
\usage{
call_that_connection(url, ...)
}
\arguments{
\item{url}{The API's URL address}

\item{...}{Available to allow backwards compatability in case more arguments are implemented in later versions of this package. Not in use today.}
}
\value{
A \code{call_that_connection} object
}
\description{
Creates a generic API connection object
}
