% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_setup.R
\name{call_that_package_check}
\alias{call_that_package_check}
\title{Runs the callthat tests for R check}
\usage{
call_that_package_check(package, reporter = testthat::ProgressReporter, ...)
}
\arguments{
\item{package}{The name of the package}

\item{reporter}{testthat reporter}
}
\description{
Runs the callthat tests for R check
}
