% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_plumber.R
\name{call_that_plumber_running}
\alias{call_that_plumber_running}
\title{Confirm if a local plumber app is running}
\usage{
call_that_plumber_running(api_connection, ...)
}
\arguments{
\item{api_connection}{A call_that_connection object}

\item{...}{Available to allow backwards compatability in case more arguments are implemented in later versions of this package. Not in use today.}
}
\description{
Confirm if a local plumber app is running
}
\details{
It returns a TRUE/FALSE based on if the R session that is running the
plumber API is running.
}
\seealso{
call_that_plumber_start
}
