% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_rsc.R
\name{call_that_rsc_connection}
\alias{call_that_rsc_connection}
\title{Creates a connection object with an RStudio Connect server}
\usage{
call_that_rsc_connection(url, key = NULL)
}
\arguments{
\item{url}{The URL address for the RStudio Connect server}

\item{key}{The personal authentication key created in the RStudio Connect server}
}
\value{
A \code{call_that_connection} \ \code{call_that_rsc_connection} object
}
\description{
Creates a connection object with an RStudio Connect server
}
