% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_session.R
\name{call_that_session_stop}
\alias{call_that_session_stop}
\title{Stops an API connection}
\usage{
call_that_session_stop(api_connection)
}
\arguments{
\item{api_connection}{A \code{call_that_connection} object.}
}
\description{
Stops an API connection
}
\details{
Use this function at the end of the tests to make sure the connection
is closed.  It is meant to make sure the R session running the local API is
stopped.
}
