% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_api.R
\name{callthat_api}
\alias{callthat_api}
\title{Generic function to make API calls}
\usage{
callthat_api(
  api_connection,
  endpoint,
  request = NULL,
  headers = list(),
  action = c("GET", "POST", "PUT"),
  ...
)
}
\arguments{
\item{api_connection}{A \code{callthat_connection} object.}

\item{endpoint}{A character variable with containing the endpoint's name.}

\item{request}{The body or query sent to the REST API.}

\item{headers}{A \code{list} object containing a named list of headers to pass to the API.}

\item{action}{The action to request from the API. Supported today are: GET, POST and PUT.}

\item{...}{Other arguments to pass to the REST API call.}
}
\description{
Generic function to make API calls
}
\details{
It provides a single interface to perform the different calls to an API, such as GET or
POST.  This function is meant to be used by developers who which to extend \code{callthat} to
integrate with other API server types.
}
