% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_api_put.R
\name{callthat_api_put}
\alias{callthat_api_put}
\title{Submits a PUT command against a REST API}
\usage{
callthat_api_put(
  api_connection,
  endpoint = "",
  body = NULL,
  headers = list(),
  ...
)
}
\arguments{
\item{api_connection}{A \code{callthat_connection} object}

\item{endpoint}{The name of the API's endpoint.  Defaults to an empty character vector.}

\item{body}{The body or query sent to the REST API.}

\item{headers}{A \code{list} object containing a named list of headers to pass to the API.}

\item{...}{Other arguments to pass to the REST API call.}
}
\value{
An \code{httr} \code{request} object
}
\description{
Submits a PUT command against a REST API
}
