% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_connection.R
\name{callthat_connection}
\alias{callthat_connection}
\title{Creates a generic API connection object}
\usage{
callthat_connection(url, ...)
}
\arguments{
\item{url}{The API's URL address}

\item{...}{Available to allow backwards compatability in case more arguments are implemented in later versions of this package. Not in use today.}
}
\value{
A \code{callthat_connection} object
}
\description{
Creates a generic API connection object
}
