% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_plumber.R
\name{callthat_plumber_stop}
\alias{callthat_plumber_stop}
\title{Stops a local plumber API}
\usage{
callthat_plumber_stop(api_connection)
}
\arguments{
\item{api_connection}{A callthat_connection object}
}
\description{
Stops a local plumber API
}
\details{
It stops the R session that is running a local API.
}
