% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callthat_session.R
\name{call_that_available_tests}
\alias{call_that_available_tests}
\title{Locates available tests in the package}
\usage{
call_that_available_tests(test_directory = "tests/testthat")
}
\arguments{
\item{test_directory}{Location of the test scripts}
}
\description{
Locates available tests in the package
}
\details{
It locates plumber tests inside the package.  It is meant to run
against the package's source folders. It looks for file names with the prefix
'test-plumber-...'.
}
